backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
whitenoise==6.6.0
python-dotenv==1.0.1
 google-cloud-storage==2.16.0
 google-cloud-firestore==2.16.1
 google-cloud-aiplatform==1.66.0
 google-cloud-speech==2.27.0
 google-cloud-texttospeech==2.17.2
 firebase-admin==6.5.0
 gunicorn==22.0.0
 uvicorn==0.30.1
 Pillow==10.3.0
 python-docx==1.1.2

